refactoring of main and database update
