+++ application/database/initalization/setup.sql
@@ -14,6 +14,7 @@
 \i database/table/skill.sql
 \i database/table/dictionary.sql
 \i database/table/term.sql
+\i database/table/schedule.sql
 
 -- Relation Tables
 \i database/table/person_action.sql
@@ -23,12 +24,10 @@
 \i database/table/person_sleep_schedule.sql
 \i database/table/person_skill.sql
 \i database/table/dictionary_term.sql
-\i database/table/quick_task.sql
-\i database/table/person_quick_task.sql
-\i database/table/schedule.sql
-\i database/table/frequency.sql
 \i database/table/schedule_weekday.sql
-
+\i database/table/intention_instance.sql
+\i database/table/person_intention_schedule.sql
+\i database/table/person_dictionary.sql
 
 
 -- Relations
