--- application/database/table/frequency.sql
@@ -1,10 +1,0 @@
-CREATE TABLE
-    "frequency"(
-        "id" uuid NOT NULL UNIQUE PRIMARY KEY,
-        "hour" integer NOT NULL,
-        "day" integer NOT NULL,
-        "week" integer NOT NUll,
-        "month" integer NOT NUll,
-        "year" integer NOT NUll,
-        "subsequent" boolean NOT NULL DEFAULT FALSE
-    );
