+++ application/database/table/intention.sql
@@ -3,10 +3,13 @@ CREATE TABLE
         "id" uuid NOT NULL PRIMARY KEY,
         "action_id" uuid NOT NULL REFERENCES "action" ("id"),
         "person_id" uuid NOT NULL REFERENCES "person" ("id"),
-        "intended_time" timestamp(0)
+        "created_on_date_and_time" timestamp(0)
+        with
+            time zone NOT NULl,
+            "intended_date_and_time" timestamp(0)
         with
             time zone NULL,
-            "complete_by" timestamp(0)
+            "complete_by_date_and_time" timestamp(0)
         with
             time zone NULL,
             "predicted_duration" interval NULL
