--- application/database/table/intention_status.sql
@@ -1,7 +1,0 @@
-CREATE TABLE "public".intention_status
-(
- "id"   uuid NOT NULL,
- Name text NOT NULL,
- CONSTRAINT PK_goal_state PRIMARY KEY ( "id" )
-);
-
