+++ application/database/table/person_dictionary.sql
@@ -1,0 +1,6 @@
+CREATE TABLE
+    "person_dictionary" (
+        "person_id" uuid NOT NULL REFERENCES "person" ("id"),
+        "dictionary_id" uuid NOT NULL REFERENCES "dictionary" ("id"),
+        PRIMARY KEY ("person_id", "dictionary_id")
+    );
