--- application/database/table/quick_task.sql
@@ -1,7 +1,0 @@
-CREATE TABLE
-    "quick_task"(
-        "id" uuid NOT NULL UNIQUE PRIMARY KEY,
-        "title" text NOT NULL,
-        "description" text NULL,
-        "completed" boolean NOT NULL DEFAULT FALSE
-    );
