+++ application/database/table/schedule.sql
@@ -1,1 +1,6 @@
-CREATE TABLE "schedule" ( "id" uuid NOT NULL UNIQUE PRIMARY KEY);
+CREATE TABLE
+    "schedule" (
+        "id" uuid NOT NULL UNIQUE PRIMARY KEY,
+        "frequency_interval" interval NOT NULL,
+        "subsequent" boolean NOT NULL DEFAULT FALSE
+    );
