+++ application/src/api.rs
@@ -2,3 +2,4 @@ pub mod guards;
 pub mod handler;
 pub mod response;
 pub mod router;
+pub mod web_server;
