+++ application/src/configuration.rs
@@ -1,0 +1,2 @@
+pub mod application;
+pub mod logging;
