+++ application/src/configuration/application.rs
@@ -1,0 +1,9 @@
+use anyhow::Result;
+
+use super::logging;
+
+pub async fn configure() -> Result<()> {
+    logging::configure_logging();
+
+    Ok(())
+}
