+++ application/src/data/query.rs
@@ -4,5 +4,4 @@ pub mod email_address;
 pub mod frequency;
 pub mod goal;
 pub mod person;
-pub mod quick_task;
 pub mod sleep_schedule;
