--- application/src/data/query/quick_task.rs
@@ -1,1 +1,0 @@
-pub mod get_quick_task_list_by_person_id;
